requirements.txt
streamlit
psycopg2-binary
python-dotenv
pandas
pillow
streamlit-option-menu
streamlit-autorefresh
statsmodels
scikit-learn
joblib
openai
